requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
python-jose[cryptography]
pydantic
email-validator
python-multipart
beanie
motor
passlib>=1.7.4
bcrypt==3.2.2
passlib[bcrypt]
cloudinary
httpx
google-auth
google-auth-oauthlib
google-auth-httplib2